refactor: clean up imports and reorganize schema definitions in auth and mfa modules
