+++ app/ldap_protocol/policies/audit/monitor.py
@@ -22,7 +22,6 @@ from ldap_protocol.auth.exceptions.mfa import (
     MFATokenError,
     NetworkPolicyError,
 )
-from ldap_protocol.auth.schemas import OAuth2Form
 from ldap_protocol.identity.exceptions import (
     AuthorizationError,
     AuthValidationError,
@@ -35,6 +34,7 @@ from ldap_protocol.kerberos.exceptions import KRBAPIChangePasswordError
 from ldap_protocol.multifactor import MFA_HTTP_Creds
 from ldap_protocol.objects import OperationEvent
 from ldap_protocol.policies.audit.audit_use_case import AuditUseCase
+from ldap_protocol.policies.audit.dataclasses import OAuth2FormDTO
 from ldap_protocol.policies.audit.events.factory import (
     RawAuditEventBuilderRedis,
 )
@@ -224,7 +224,7 @@ class AuditMonitorUseCase:
     def wrap_login(self, attr: _T) -> _T:
         @wraps(attr)
         async def wrapped_login(
-            form: OAuth2Form,
+            form: OAuth2FormDTO,
             url: URL,
             ip: IPv4Address | IPv6Address,
             user_agent: str,
